+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- feat: add person and group properties for feature flags ([#366](https://github.com/PostHog/posthog-ios/pull/366))
+- feat: add person and group properties for feature flags ([#373](https://github.com/PostHog/posthog-ios/pull/373))
 
 ## 3.29.0 - 2025-07-15